requirements.txt
selenium
selenium-wire
webdriver_manager
pillow